requirements.txt
Django==3.2
python-webpack-boilerplate
djangorestframework==3.12.4
django-push-notifications==2.0.0
py-vapid==1.8.2

psycopg2-binary

